Added Bullet class and animated background
Added Bullet class which is a child of the Entity class. Inside are some template bullet shot types that'll be used ingame (most likely with a few tweaks).

Added CSS to animate the background image on the Canvas when the game starts.

Modified and added more functions to functions.js