Final additions and changes
Decided on 'Vector Space' for the name of the game.

Added title screen 'animation'.

Added 'levels', powerups, score and winning end condition to the game.

Added more sounds to use in and out of game.

Added more sprites to use for variety.

Altered the way the canvas background 'animation' works.

Altered the Credits.

Renamed variables for clarity.

Added enemy 'behaviors' to allow for specified movement and shooting patterns.

Removed an unused ShotType function.